Update for 5.4.9
- Handle upgrade prompts (this installer really isn't that silent is it)

+++ unifi-controller/tools/chocolateyuninstall.ps1
@@ -43,17 +43,20 @@ $toolsPath = Split-Path $MyInvocation.MyCommand.Definition
 $ahkScript = "$toolsPath\unifi-uninstall.ahk"
 AutoHotkey $ahkScript $packageArgs.fileFullPath
 
-Remove-NetFirewallRule -Name UniFi-Mgmt-In 
-Remove-NetFirewallRule -Name UniFi-Mgmt-Out
-Remove-NetFirewallRule -Name UniFi-DvcInfrm-In
-Remove-NetFirewallRule -Name UniFi-DvcInfrm-Out
-Remove-NetFirewallRule -Name UniFi-Ctrlr-In
-Remove-NetFirewallRule -Name UniFi-Ctrlr-Out
-Remove-NetFirewallRule -Name UniFi-PrtlRdr-In
-Remove-NetFirewallRule -Name UniFi-PrtlRdr-Out
-Remove-NetFirewallRule -Name UniFi-PrtlRdrSsl-In
-Remove-NetFirewallRule -Name UniFi-PrtlRdrSsl-Out
-#Remove-NetFirewallRule -Name UniFi-DB-In
-#Remove-NetFirewallRule -Name UniFi-DB-Out
-Remove-NetFirewallRule -Name UniFi-DvcDisc-In
-Remove-NetFirewallRule -Name UniFi-DvcDisc-Out
+# Win8/2008 supports 'Remove-NetFirewallRule' - https://technet.microsoft.com/en-us/library/jj554908(v=wps.620).aspx
+if ([Environment]::OSVersion.Version.Major -ge 6) {
+    Remove-NetFirewallRule -Name UniFi-Mgmt-In 
+    Remove-NetFirewallRule -Name UniFi-Mgmt-Out
+    Remove-NetFirewallRule -Name UniFi-DvcInfrm-In
+    Remove-NetFirewallRule -Name UniFi-DvcInfrm-Out
+    Remove-NetFirewallRule -Name UniFi-Ctrlr-In
+    Remove-NetFirewallRule -Name UniFi-Ctrlr-Out
+    Remove-NetFirewallRule -Name UniFi-PrtlRdr-In
+    Remove-NetFirewallRule -Name UniFi-PrtlRdr-Out
+    Remove-NetFirewallRule -Name UniFi-PrtlRdrSsl-In
+    Remove-NetFirewallRule -Name UniFi-PrtlRdrSsl-Out
+    #Remove-NetFirewallRule -Name UniFi-DB-In
+    #Remove-NetFirewallRule -Name UniFi-DB-Out
+    Remove-NetFirewallRule -Name UniFi-DvcDisc-In
+    Remove-NetFirewallRule -Name UniFi-DvcDisc-Out
+}
